int to Long pour les iDs+ sauvegarde frigo

+++ src/main/java/com/takima/chefkit/DTO/ingredientsDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 @Builder
 public class ingredientsDTO {
-    private int idIngredient;
+    private Long idIngredient;
     private String nomIngredient;
     private String categorie;
 }
